+++ src/Bootstrap.php
@@ -9,6 +9,7 @@ use davidhirtz\yii2\cms\models\queries\EntryQuery;
 use davidhirtz\yii2\cms\modules\admin\data\EntryActiveDataProvider;
 use davidhirtz\yii2\cms\modules\admin\widgets\forms\fields\EntryParentIdSelectField;
 use davidhirtz\yii2\cms\modules\admin\widgets\grids\EntryGridView;
+use davidhirtz\yii2\cms\modules\admin\widgets\grids\SectionParentEntryGridView;
 use davidhirtz\yii2\cms\tenant\behaviors\EntryTenantBehavior;
 use davidhirtz\yii2\cms\tenant\behaviors\TenantEntryBehavior;
 use davidhirtz\yii2\skeleton\filters\PageCache;
@@ -40,6 +41,7 @@ class Bootstrap implements BootstrapInterface
             EntryQuery::class => models\queries\EntryQuery::class,
             TenantGridView::class => modules\admin\widgets\grids\TenantGridView::class,
             PageCache::class => filters\PageCache::class,
+            SectionParentEntryGridView::class => widgets\grids\SectionParentEntryGridView::class,
         ];
 
         foreach ($definitions as $oldClass => $newClass) {