+++ src/modules/admin/widgets/grids/SectionParentEntryGridView.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace davidhirtz\yii2\cms\tenant\widgets\grids;
+
+use davidhirtz\yii2\cms\tenant\widgets\grids\traits\EntryGridViewTrait;
+
+class SectionParentEntryGridView extends \davidhirtz\yii2\cms\modules\admin\widgets\grids\SectionParentEntryGridView
+{
+    use EntryGridViewTrait;
+}