+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 - Removed `TenantIdFieldBehavior`
 - Renamed `TenantDropdownTrait` and `EntryCountColumnTrait`
 
+# 1.2.3 (Dec 6, 2025)
+
+- Fixed `TenantEntryBehavior::onBeforeDelete()` to prevent deletion of a tenant that still has entries
+
+# 1.2.2 (Nov 24, 2025)
+
+- Added `EntryGridViewTrait` and `SectionParentEntryGridView`
+
+# 1.2.1 (Jul 15, 2025)
+
+- Added slug index creation
+
 # 1.2.0 (May 26, 2025)
 
 - Requires PHP 8.3+